apis/ga4-campaign/requirements.txt
# GA4 Campaign Extractor API - Dependencies
# Python 3.12+

# Web Framework
flask>=3.0.0
gunicorn>=22.0.0

# Google Cloud
google-cloud-bigquery>=3.14.0
google-analytics-data>=0.18.0
google-auth>=2.27.0
google-cloud-secret-manager>=2.18.0

# Data Processing
pytz>=2023.3

# Utilities
python-dateutil>=2.8.2
